Create Rod_cutting.cpp
Task description: Given a rod of length b inches and a table of prices ai for i=1 to n (ai is the price of a rod of length i), determine the maximum revenue rn obtainable by cutting up the rod and selling the pieces. Note that if the price ab for a rod of length b is large enough, an optimal solution may require no cutting at all.